エラーの解消
一部コメントの削除
一部パラメーターの変更

+++ judge/Len_judge.h
@@ -10,7 +10,7 @@ class Len_judge : public Judge
 {
     public:
         bool judge();
-        void setLength(int setLen);
+        void setLength(double setLen);
         double setLen;
         double slen;
